+++ tdmclient/tools/repl.py
@@ -21,7 +21,7 @@ except ModuleNotFoundError:
 
 
 def help():
-    print("""Usage: python3 -m tdmclient.tools.repl [options]
+    print(f"""Usage: python3 -m tdmclient.tools.repl [options]
 Read-eval-print loop with link and synchronization with Thymio
 
 Options:
@@ -30,7 +30,7 @@ Options:
   --robotid=I    robot id; default=any
   --robotname=N  robot name; default=any
   --tdmaddr=H    tdm address (default: localhost or from zeroconf)
-  --tdmport=P    tdm port (default: from zeroconf)
+  --tdmport=P    tdm port or "default" for {ClientAsync.DEFAULT_TDM_PORT} (default: from zeroconf)
 """)
 
 
@@ -69,7 +69,7 @@ def main(argv=None):
             elif arg == "--tdmaddr":
                 tdm_addr = val
             elif arg == "--tdmport":
-                tdm_port = int(val)
+                tdm_port = ClientAsync.DEFAULT_TDM_PORT if val == "default" else int(val)
 
     with ClientAsync(tdm_addr=tdm_addr, tdm_port=tdm_port,
                      password=password) as client: